+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * The user list is back! It now only returns 25 users at a time, but you can retrieve any page
   It accepts sorting by kudos or by account age. Kudos is sorted descending. Account age sorted ascending.
-  
+* Now model list should display accurate performance average per model
+
 # 4.6.8
 
 * Reduced kudos calculation for unapproved textgen models a bit.